backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
pydantic-settings==2.6.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mne==1.8.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
google-generativeai==0.8.3
python-dotenv==1.0.1
aiofiles==24.1.0
pypdf==5.1.0
matplotlib==3.9.2
scipy==1.14.1
httpx==0.27.2
pytest==8.3.3
Pillow>=10.0.0
